keep sysutils.lib.sa too

+++ scripts/python/make-dist.py
@@ -1,5 +1,5 @@
 #! /usr/bin/env python
-# $Id: make-dist.py,v 1.30 2008-04-03 15:48:55 jkrell Exp $
+# $Id: make-dist.py,v 1.31 2008-04-03 15:50:17 jkrell Exp $
 
 import sys
 import os.path
@@ -71,10 +71,12 @@ def MakeArchive(PackageSetName, Command, Extension):
                     or Extension == ".sa
                     ):
                 #
-                # Keep libm3.lib.sa and m3core.lib.sa for bootstrapping of cm3.
+                # Keep libm3.lib.sa, m3core.lib.sa, sysutils.lib.sa for bootstrapping of cm3.
                 # This check is loose to allow for multiple naming schemes.
                 #
-                if (Name.find("m3core") == -1) and (Name.find("libm3") == -1):
+                if ((Name.find("m3core") == -1)
+                        and (Name.find("libm3") == -1):
+                        and (Name.find("sysutils") == -1)):
                     DeleteFile(os.path.join(Directory, Name))
 
     CreateDirectory(SymbolsRoot)
